test: consolidate FP8 GDN accuracy coverage
